Move object-specific code out of `maybe_partial`
