+++ crates/fj-kernel/src/partial/maybe_partial.rs
@@ -1,13 +1,7 @@
-use fj_math::Point;
-
 use crate::{
-    geometry::{path::SurfacePath, surface::SurfaceGeometry},
     get::Get,
     insert::Insert,
-    objects::{
-        Curve, Cycle, GlobalCurve, GlobalEdge, GlobalVertex, HalfEdge, Objects,
-        Surface, SurfaceVertex, Vertex,
-    },
+    objects::Objects,
     services::Service,
     storage::Handle,
     validate::{Validate, ValidationError},
@@ -156,152 +150,3 @@ where
 
 // Unfortunately, we can't add a blanket implementation from `T::Partial` for
 // `MaybePartial<T>`, as that would conflict.
-
-impl MaybePartial<Curve> {
-    /// Access the path
-    pub fn path(&self) -> Option<SurfacePath> {
-        match self {
-            Self::Full(full) => Some(full.path()),
-            Self::Partial(partial) => partial.path,
-        }
-    }
-
-    /// Access the surface
-    pub fn surface(&self) -> Option<Handle<Surface>> {
-        match self {
-            Self::Full(full) => Some(full.surface().clone()),
-            Self::Partial(partial) => partial.surface.clone(),
-        }
-    }
-
-    /// Access the global form
-    pub fn global_form(&self) -> MaybePartial<GlobalCurve> {
-        match self {
-            Self::Full(full) => full.global_form().clone().into(),
-            Self::Partial(partial) => partial.global_form.clone(),
-        }
-    }
-}
-
-impl MaybePartial<Cycle> {
-    /// Access the surface
-    pub fn surface(&self) -> Option<Handle<Surface>> {
-        match self {
-            Self::Full(full) => full.surface().clone().into(),
-            Self::Partial(partial) => partial.surface(),
-        }
-    }
-}
-
-impl MaybePartial<GlobalEdge> {
-    /// Access the curve
-    pub fn curve(&self) -> MaybePartial<GlobalCurve> {
-        match self {
-            Self::Full(full) => full.curve().clone().into(),
-            Self::Partial(partial) => partial.curve.clone(),
-        }
-    }
-
-    /// Access the vertices
-    pub fn vertices(&self) -> [MaybePartial<GlobalVertex>; 2] {
-        match self {
-            Self::Full(full) => {
-                full.vertices().access_in_normalized_order().map(Into::into)
-            }
-            Self::Partial(partial) => partial.vertices.clone(),
-        }
-    }
-}
-
-impl MaybePartial<GlobalVertex> {
-    /// Access the position
-    pub fn position(&self) -> Option<Point<3>> {
-        match self {
-            Self::Full(full) => Some(full.position()),
-            Self::Partial(partial) => partial.position,
-        }
-    }
-}
-
-impl MaybePartial<HalfEdge> {
-    /// Access the curve
-    pub fn curve(&self) -> MaybePartial<Curve> {
-        match self {
-            Self::Full(full) => full.curve().clone().into(),
-            Self::Partial(partial) => partial.curve(),
-        }
-    }
-
-    /// Access the front vertex
-    pub fn front(&self) -> MaybePartial<Vertex> {
-        match self {
-            Self::Full(full) => full.front().clone().into(),
-            Self::Partial(partial) => {
-                let [_, front] = &partial.vertices;
-                front.clone()
-            }
-        }
-    }
-
-    /// Access the vertices
-    pub fn vertices(&self) -> [MaybePartial<Vertex>; 2] {
-        match self {
-            Self::Full(full) => full.vertices().clone().map(Into::into),
-            Self::Partial(partial) => partial.vertices.clone(),
-        }
-    }
-}
-
-impl MaybePartial<Surface> {
-    /// Access the geometry
-    pub fn geometry(&self) -> Option<SurfaceGeometry> {
-        match self {
-            Self::Full(full) => Some(full.geometry()),
-            Self::Partial(partial) => partial.geometry,
-        }
-    }
-}
-
-impl MaybePartial<SurfaceVertex> {
-    /// Access the position
-    pub fn position(&self) -> Option<Point<2>> {
-        match self {
-            Self::Full(full) => Some(full.position()),
-            Self::Partial(partial) => partial.position,
-        }
-    }
-
-    /// Access the surface
-    pub fn surface(&self) -> Option<Handle<Surface>> {
-        match self {
-            Self::Full(full) => Some(full.surface().clone()),
-            Self::Partial(partial) => partial.surface.clone(),
-        }
-    }
-
-    /// Access the global form
-    pub fn global_form(&self) -> MaybePartial<GlobalVertex> {
-        match self {
-            Self::Full(full) => full.global_form().clone().into(),
-            Self::Partial(partial) => partial.global_form.clone(),
-        }
-    }
-}
-
-impl MaybePartial<Vertex> {
-    /// Access the curve
-    pub fn curve(&self) -> MaybePartial<Curve> {
-        match self {
-            Self::Full(full) => full.curve().clone().into(),
-            Self::Partial(partial) => partial.curve.clone(),
-        }
-    }
-
-    /// Access the surface form
-    pub fn surface_form(&self) -> MaybePartial<SurfaceVertex> {
-        match self {
-            Self::Full(full) => full.surface_form().clone().into(),
-            Self::Partial(partial) => partial.surface_form.clone(),
-        }
-    }
-}
